moved bp-rp lookup to star class
